+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/rnw-community/rnw-community/compare/v2.13.0...v2.14.0) (2026-08-18)
+
+### Features
+
+- **react-native-collapsible-header:** reduced-motion-aware snap and generic scrollable support ([1a81d93](https://github.com/rnw-community/rnw-community/commit/1a81d936f581ae3aa8fd2e8c4900ee5b3c48d0bc))
+
 # [2.13.0](https://github.com/rnw-community/rnw-community/compare/v2.12.13...v2.13.0) (2026-08-18)
 
 ### Features