+++ packages/react-native-collapsible-header-example/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/rnw-community/rnw-community/compare/v2.13.0...v2.14.0) (2026-08-18)
+
+**Note:** Version bump only for package @rnw-community/react-native-collapsible-header-example
+
 # [2.13.0](https://github.com/rnw-community/rnw-community/compare/v2.12.13...v2.13.0) (2026-08-18)
 
 ### Features